refactored breathing menu
