+++ public/service-worker.js
@@ -37,7 +37,7 @@
 /* eslint-disable indent, no-unused-vars, no-multiple-empty-lines, max-nested-callbacks, space-before-function-paren, quotes, comma-spacing */
 'use strict';
 
-var precacheConfig = [["08e7f56bfde742d0901b84e2485eb123.svg","08e7f56bfde742d0901b84e2485eb123"],["0bffba416891219267efcf4e59c3c015.svg","0bffba416891219267efcf4e59c3c015"],["2751ee43015f9884c3642f103b7f70c9.woff2","2751ee43015f9884c3642f103b7f70c9"],["30799efa5bf74129468ad4e257551dc3.eot","30799efa5bf74129468ad4e257551dc3"],["372f6afd60f19c9281a5d36b9bc9b2c1.svg","372f6afd60f19c9281a5d36b9bc9b2c1"],["39b2c3031be6b4ea96e2e3e95d307814.woff2","39b2c3031be6b4ea96e2e3e95d307814"],["3b813c2ae0d04909a33a18d792912ee7.woff","3b813c2ae0d04909a33a18d792912ee7"],["3c546bf7e955aeaf180bca3d2743e204.svg","3c546bf7e955aeaf180bca3d2743e204"],["46e48ce0628835f68a7369d0254e4283.ttf","46e48ce0628835f68a7369d0254e4283"],["4d9f3f9e5195e7b074bb63ba4ce42208.eot","4d9f3f9e5195e7b074bb63ba4ce42208"],["574fd0b50367f886d359e8264938fc37.woff2","574fd0b50367f886d359e8264938fc37"],["69f8a0617ac472f78e45841323a3df9e.woff2","69f8a0617ac472f78e45841323a3df9e"],["7500519de3d82e33d1587f8042e2afcb.woff","7500519de3d82e33d1587f8042e2afcb"],["7ad07142b51dd9b7c1b6347603383c6a.svg","7ad07142b51dd9b7c1b6347603383c6a"],["7ef525e40bfdaf62100e47dd4be50351.svg","7ef525e40bfdaf62100e47dd4be50351"],["7ef9009b51f2e3ced613809250f98a30.svg","7ef9009b51f2e3ced613809250f98a30"],["894a2ede85a483bf9bedefd4db45cdb9.ttf","894a2ede85a483bf9bedefd4db45cdb9"],["94998475f6aea65f558494802416c1cf.ttf","94998475f6aea65f558494802416c1cf"],["954bbdeb86483e4ffea00c4591530ece.woff2","954bbdeb86483e4ffea00c4591530ece"],["a0e808d38c78ec1022d2f3eae5c5d980.svg","a0e808d38c78ec1022d2f3eae5c5d980"],["a858b276f999917b0698761ffcf428aa.gif","a858b276f999917b0698761ffcf428aa"],["a990f611f2305dc12965f186c2ef2690.eot","a990f611f2305dc12965f186c2ef2690"],["ba3dcd8903e3d0af5de7792777f8ae0d.woff","ba3dcd8903e3d0af5de7792777f8ae0d"],["bc540e50b175a23a39a1b3bffc6e0785.png","bc540e50b175a23a39a1b3bffc6e0785"],["breathe.appcache","a364b04a9f427cd35a11efb5975c9dee"],["bundle.css","67084c97a77fdf3ed72d224dda3d2e96"],["bundle.js","66348cf2d9f62b78bc7777c6636ef5d7"],["catLanding.png","62cd9ffc885778eb42f670405c330ad2"],["dc81817def276b4f21395f7ea5e88dcd.woff","dc81817def276b4f21395f7ea5e88dcd"],["df7b648ce5356ea1ebce435b3459fd60.ttf","df7b648ce5356ea1ebce435b3459fd60"],["dfe56a876d0282555d1e2458e278060f.eot","dfe56a876d0282555d1e2458e278060f"],["e31fcf1885e371e19f5786c2bdfeae1b.ttf","e31fcf1885e371e19f5786c2bdfeae1b"],["ea463d8bf1904b3cece6052283b04971.svg","ea463d8bf1904b3cece6052283b04971"],["ecdd509cadbf1ea78b8d2e31ec52328c.eot","ecdd509cadbf1ea78b8d2e31ec52328c"],["f079d603854bcd795eaabf2ef7ae723b.svg","f079d603854bcd795eaabf2ef7ae723b"],["fc78759e93a6cac50458610e3d9d63a0.woff","fc78759e93a6cac50458610e3d9d63a0"],["index.html","8f31561422e317c2d60cc252519cdb41"],["manifest.json","56d9b70e3f1dfb4959f54d83a961979e"],["music.mp3","61d004e08a8c65b82ec8ea3cea23ce38"]];
+var precacheConfig = [["08e7f56bfde742d0901b84e2485eb123.svg","08e7f56bfde742d0901b84e2485eb123"],["0bffba416891219267efcf4e59c3c015.svg","0bffba416891219267efcf4e59c3c015"],["2751ee43015f9884c3642f103b7f70c9.woff2","2751ee43015f9884c3642f103b7f70c9"],["30799efa5bf74129468ad4e257551dc3.eot","30799efa5bf74129468ad4e257551dc3"],["372f6afd60f19c9281a5d36b9bc9b2c1.svg","372f6afd60f19c9281a5d36b9bc9b2c1"],["39b2c3031be6b4ea96e2e3e95d307814.woff2","39b2c3031be6b4ea96e2e3e95d307814"],["3b813c2ae0d04909a33a18d792912ee7.woff","3b813c2ae0d04909a33a18d792912ee7"],["3c546bf7e955aeaf180bca3d2743e204.svg","3c546bf7e955aeaf180bca3d2743e204"],["46e48ce0628835f68a7369d0254e4283.ttf","46e48ce0628835f68a7369d0254e4283"],["4d9f3f9e5195e7b074bb63ba4ce42208.eot","4d9f3f9e5195e7b074bb63ba4ce42208"],["574fd0b50367f886d359e8264938fc37.woff2","574fd0b50367f886d359e8264938fc37"],["69f8a0617ac472f78e45841323a3df9e.woff2","69f8a0617ac472f78e45841323a3df9e"],["7500519de3d82e33d1587f8042e2afcb.woff","7500519de3d82e33d1587f8042e2afcb"],["7ad07142b51dd9b7c1b6347603383c6a.svg","7ad07142b51dd9b7c1b6347603383c6a"],["7ef525e40bfdaf62100e47dd4be50351.svg","7ef525e40bfdaf62100e47dd4be50351"],["7ef9009b51f2e3ced613809250f98a30.svg","7ef9009b51f2e3ced613809250f98a30"],["894a2ede85a483bf9bedefd4db45cdb9.ttf","894a2ede85a483bf9bedefd4db45cdb9"],["94998475f6aea65f558494802416c1cf.ttf","94998475f6aea65f558494802416c1cf"],["954bbdeb86483e4ffea00c4591530ece.woff2","954bbdeb86483e4ffea00c4591530ece"],["a0e808d38c78ec1022d2f3eae5c5d980.svg","a0e808d38c78ec1022d2f3eae5c5d980"],["a858b276f999917b0698761ffcf428aa.gif","a858b276f999917b0698761ffcf428aa"],["a990f611f2305dc12965f186c2ef2690.eot","a990f611f2305dc12965f186c2ef2690"],["ba3dcd8903e3d0af5de7792777f8ae0d.woff","ba3dcd8903e3d0af5de7792777f8ae0d"],["bc540e50b175a23a39a1b3bffc6e0785.png","bc540e50b175a23a39a1b3bffc6e0785"],["breathe.appcache","96c6895ec9b6930028c72983f05f8a32"],["bundle.css","67084c97a77fdf3ed72d224dda3d2e96"],["bundle.js","e8c38ad38045afa4987072379a08873a"],["catLanding.png","62cd9ffc885778eb42f670405c330ad2"],["dc81817def276b4f21395f7ea5e88dcd.woff","dc81817def276b4f21395f7ea5e88dcd"],["df7b648ce5356ea1ebce435b3459fd60.ttf","df7b648ce5356ea1ebce435b3459fd60"],["dfe56a876d0282555d1e2458e278060f.eot","dfe56a876d0282555d1e2458e278060f"],["e31fcf1885e371e19f5786c2bdfeae1b.ttf","e31fcf1885e371e19f5786c2bdfeae1b"],["ea463d8bf1904b3cece6052283b04971.svg","ea463d8bf1904b3cece6052283b04971"],["ecdd509cadbf1ea78b8d2e31ec52328c.eot","ecdd509cadbf1ea78b8d2e31ec52328c"],["f079d603854bcd795eaabf2ef7ae723b.svg","f079d603854bcd795eaabf2ef7ae723b"],["fc78759e93a6cac50458610e3d9d63a0.woff","fc78759e93a6cac50458610e3d9d63a0"],["index.html","8f31561422e317c2d60cc252519cdb41"],["manifest.json","56d9b70e3f1dfb4959f54d83a961979e"],["music.mp3","61d004e08a8c65b82ec8ea3cea23ce38"]];
 var cacheName = 'sw-precache-v2-pages-cache-v1-' + (self.registration ? self.registration.scope : '');
